requirements.txt
streamlit>=1.36
pillow>=10.0
tqdm>=4.66
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
zstandard>=0.22
pytest>=8.0
